// src/services/tealium/event.enum.ts
export enum TealiumEventEnum {
  /**
   * User login
   */
  userLogin = 'user_login',

  /**
   * User logout
   */
  userLogout = 'user_logout',

  /**
   * User register
   */
  userRegister = 'user_register',

  /**
   * User sign-up
   */
  emailSignup = 'email_signup',

  /**
   * Call now
   */
  callNow = 'call_now',

  /**
   * Email now
   */
  emailNow = 'email_now',

  /**
   * App download
   */
  appDownload = 'app_download',

  /**
   * Save property
   */
  saveProperty = 'save_property',

  /**
   * Call back
   */
  callBack = 'call_back',

  /**
   * Whatsapp
   */
  whatsappNow = 'whatsapp_now',

  /**
   * SMS
   */
  smsNow = 'sms_now',

  /**
   * Default view event
   */
  default = 'view',

  /**
   * Search view
   */
  search = 'search_view',

  /**
   * Phone reveal
   */
  phoneReveal = 'phone_reveal',

  /**
   * open gallery modal
   */
  openGalleryModal = 'open_gallery',

  /**
   * close gallery modal
   */
  closeGalleryModal = 'close_gallery',

  /**
   * scroll gallery modal
   */
  scrollGalleryModal = 'gallery_scroll',
}