// src/services/tealium/event-action.enum.ts
export enum TealiumEventActionEnum {
  /**
   * User sign in
   */
  signIn = 'sign in',

  /**
   * User sign out
   */
  signOut = 'sign out',

  /**
   * User register
   */
  register = 'register',

  /**
   * Email newsletter
   */
  emailAlert = 'email alert',

  /**
   * find action
   */
  find = 'find',

  /**
   * Call
   */
  call = 'call',

  /**
   * Email
   */
  email = 'email',

  /**
   * Travel time
   */
  travelTime = 'travel Time',

  /**
   * App
   */
  app = 'app',

  /**
   * Mortgage Calculator
   */
  mortgageCalculator = 'mortgage calculator',

  /**
   * Saved Properties
   */
  savedProperties = 'saved properties',

  /**
   * Whatsapp
   */
  whatsapp = 'whatsapp',

  /**
   * sms
   */
  sms = 'sms',

  /**
   * map - amenity selection
   */
  mapAmenitySelection = 'amenity_selected',

  /**
   * map - commute time search open
   */
  mapCommuteTimeSearchOpen = 'search_open',
}