// src/services/tealium/event-label.enum.ts
export enum TealiumEventLabelEnum {
  /**
   * sign up
   */
  signUp = 'sign up',

  /**
   * call directly
   */
  callDirect = 'call direct',

  /**
   * call back
   */
  callBack = 'call back',

  /**
   * email
   */
  email = 'email',

  /**
   * email agent
   */
  emailNow = 'email now',

  /**
   * calculate
   */
  calculate = 'calculate',

  /**
   * download
   */
  download = 'download',

  /**
   * save
   */
  save = 'save',

  /**
   * phone reveal
   */
  phoneReveal = 'phone reveal',
}